Handle FlatBuffers in-process with FlatSharp, and add more flatbuffer editors. (#166)
Thanks @sora10pls for double checking some editors, and @Lusamine for some field identification.

Closes #115 -- tested the pokedex after randomization and it no longer crashes when randomizing+saving with the latest code in this pull request.

Adds:
* Shop editor (SWSH + LGPE)
* Move editor
* Item editor
* Symbol (Overworld Pokemon) Behavior editor
* Default Rental team editor

+++ pkNX.WinForms/Controls/EncounterList8.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Windows.Forms;
 using pkNX.Sprites;
-using pkNX.Structures;
+using pkNX.Structures.FlatBuffers;
 
 namespace pkNX.WinForms
 {
@@ -115,9 +115,9 @@ namespace pkNX.WinForms
         {
             int sp = Array.IndexOf(species, dgv.Rows[row].Cells[1].Value ?? species[0]);
             string formstr = (dgv.Rows[row].Cells[2].Value ?? 0).ToString();
-            int.TryParse(formstr, out var form);
+            byte.TryParse(formstr, out var form);
             string probstr = (dgv.Rows[row].Cells[3].Value ?? 0).ToString();
-            int.TryParse(probstr, out var prob);
+            byte.TryParse(probstr, out var prob);
 
             if (sp == 0)
             {
